requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.22.4
Pillow>=9.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
pandas>=1.4.0
scikit-learn>=1.0.0
segmentation-models-pytorch>=0.3.0
opencv-python>=4.6.0
